work on invoice
