+++ pages/admin/addinvoice.tsx
@@ -33,6 +33,7 @@ import { Button, OutlinedInput } from "@mui/material";
 import Paper from "@mui/material/Paper";
 import AddCustomer from "../customer/addNewCustomer";
 import AddItem from "./additem";
+import invoicesService from "../services/invoiceService";
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   "& .MuiDialogContent-root": {
@@ -331,20 +332,15 @@ export default function Guardians() {
       router.push("/admin/invoices");
     }, 1000);
 
-    await axios({
-      method: "POST",
-      url: `${api_url}/createInvoice`,
-      data: requestedData,
-      headers: {
-        "content-type": "multipart/form-data",
-      },
-    })
+    await invoicesService
+      .createInvoice(requestedData)
       .then((res) => {
         if (!res) {
           reset();
         }
       })
       .catch((err) => {});
+
     // window.location.replace("/admin/invoices");
   };
   const getItem = async () => {
